Refactor: move http/request package to internal directory
